refactor(api): always enable Anthropic API and drop litellm shim
Inline the LiteLLM adapter loader into api_anthropic.py and remove the
--enable_anthropic_api flag so /v1/messages is mounted unconditionally.
